requirements.txt
protobuf==7.34.1
selenium==4.43.0
webdriver_manager==4.0.2
google-cloud-storage
curl_cffi==0.15.0
loguru==0.7.3
langchain-text-splitters==1.1.2
tenacity==9.1.4
pymupdf==1.27.2.3
pymupdf4llm==1.27.2.3
pandas==3.0.2
beautifulsoup4==4.14.3
pytesseract==0.3.13
pillow==12.2.0
openai
qdrant-client
python-dotenv